services/ml_service/requirements.txt
fastapi==0.115.2
pandas==2.2.3
requests==2.32.3
uvicorn==0.32.0
pydantic-settings==2.5.2
pyyaml==6.0.1
jinja2==3.1.4
pyarrow==17.0.0
httpx==0.27.2
prometheus-fastapi-instrumentator==7.0.0
aioredis==2.0.1
fastapi_limiter==0.1.6